requirements-deepgram.txt
# Deepgram Live Audio Transcription Dependencies
# Install with: pip install -r requirements-deepgram.txt

# Core Deepgram SDK
deepgram-sdk>=4.8.1

# Audio processing dependencies (if not already installed)
pyaudiowpatch>=0.2.12.4
numpy>=1.21.0
scipy>=1.7.0
noisereduce>=3.0.0

# Async support
asyncio-mqtt>=0.11.0
